mysite/requirements.txt
qrcode==6.1
Django==2.2.5
requests==2.25.1
djangorestframework==3.12.4